fix(whatsapp): narrow Baileys runtime foundation
