+++ backend/app/routes/channels.py
@@ -9,7 +9,7 @@ from app.routes.channel_routers import (
     public,
     telegram,
     whatsapp,
-    whatsapp_application,
+    whatsapp_sidecar,
 )
 
 router = APIRouter(tags=["channels"])
@@ -21,7 +21,7 @@ router.include_router(public.router)
 
 # Provider routers own the SDK-compatible and webhook surfaces under
 # /v1/channels/{provider}/*.
-router.include_router(whatsapp_application.router)
+router.include_router(whatsapp_sidecar.router)
 router.include_router(whatsapp.router)
 router.include_router(telegram.router)
 router.include_router(imessage.router)
