+++ packages/whatsapp-baileys-sidecar/package.json
@@ -15,11 +15,13 @@
 	},
 	"dependencies": {
 		"baileys": "npm:@whiskeysockets/baileys@7.0.0-rc13",
-		"pino": "^10.3.1"
+		"pino": "^10.3.1",
+		"qrcode-terminal": "^0.12.0"
 	},
 	"devDependencies": {
 		"@types/bun": "^1.3.13",
 		"@types/node": "^22.20.0",
+		"@types/qrcode-terminal": "^0.12.2",
 		"typescript": "catalog:"
 	}
 }
