+++ packages/whatsapp-baileys-sidecar/src/jid.ts
@@ -1,14 +1,14 @@
 import { isJidGroup, isLidUser, isPnUser, jidDecode, jidNormalizedUser } from "baileys";
 
-export function normalizeApplicationUserJid(value: string): string | undefined {
-	return normalizeApplicationJid(value, false);
+export function normalizeUserJid(value: string): string | undefined {
+	return normalizeJid(value, false);
 }
 
-export function normalizeApplicationChatJid(value: string): string | undefined {
-	return normalizeApplicationJid(value, true);
+export function normalizeChatJid(value: string): string | undefined {
+	return normalizeJid(value, true);
 }
 
-function normalizeApplicationJid(value: string, allowGroup: boolean): string | undefined {
+function normalizeJid(value: string, allowGroup: boolean): string | undefined {
 	const decoded = jidDecode(value);
 	const normalized = jidNormalizedUser(value);
 	if (!decoded?.user || normalized !== value) return undefined;
