+++ packages/whatsapp-baileys-sidecar/src/types.ts
@@ -12,6 +12,10 @@ export type RuntimeHealth = {
 	};
 	lastDisconnectReason?: string;
 	pendingCallbackEvents?: number;
+	pairing?: {
+		method: "qr" | "code";
+		value: string;
+	};
 };
 
 export type RelayMessageRequest = {
